+++ plugins/jsx_preact.ts
@@ -63,7 +63,7 @@ export class PreactJsxEngine implements Engine {
     return element;
   }
 
-  renderSync(content: unknown, data: Data = {}): string {
+  renderSync(content: unknown, data: Data = {}) {
     const element = typeof content === "function"
       ? content(data, this.helpers)
       : content;