+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 
 ### Fixed
 - Return type in Preact engine [#403].
+- Updated dependencies: `std`, `terser`, `sass`, `esbuild`, `liquid`.
 
 ## [1.16.0] - 2023-03-21
 ### Added