+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/),
 but not always is possible (due the use of unstable features from Deno).
 Any BREAKING CHANGE between minor versions will be documented here in upper case.
 
-## [1.16.1] - Unreleased
+## [1.16.1] - 2023-03-29
 ### Added
 - `esm` option to `esbuild` plugin [#400], [#401].
 - `query` and `sort` arguments to `nav.menu()` and `nav.breadcrumb()`.
@@ -2154,7 +2154,7 @@ The first version.
 [#401]: https://github.com/lumeland/lume/issues/401
 [#403]: https://github.com/lumeland/lume/issues/403
 
-[1.16.1]: https://github.com/lumeland/lume/compare/v1.16.0...HEAD
+[1.16.1]: https://github.com/lumeland/lume/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/lumeland/lume/compare/v1.15.3...v1.16.0
 [1.15.3]: https://github.com/lumeland/lume/compare/v1.15.2...v1.15.3
 [1.15.2]: https://github.com/lumeland/lume/compare/v1.15.1...v1.15.2